requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pyjwt==2.8.0
requests==2.32.3
datetime==5.5
cryptography==42.0.8